Move personal agent DB adapter to session bridge
